Pull XSLTTransformer to own class + add failing test
