+++ org.hl7.fhir.publisher.core/src/main/java/org/hl7/fhir/igtools/publisher/Publisher.java
@@ -8707,7 +8707,7 @@ public class Publisher implements IWorkerContext.ILoggingService, IReferenceReso
           f.getOutputNames().add(dst);
           Utilities.createDirectory(dst);
         } else
-          checkMakeFile(transform(f.getSource(), f.getXslt()), dst, f.getOutputNames());
+          checkMakeFile(new XSLTTransformer(debug).transform(f.getSource(), f.getXslt()), dst, f.getOutputNames());
       } catch (Exception e) {
         log("Exception generating xslt page "+dst+" for "+f.getRelativePath()+" in "+tempDir+": "+e.getMessage());
       }
@@ -9187,48 +9187,6 @@ public class Publisher implements IWorkerContext.ILoggingService, IReferenceReso
     if (r instanceof CanonicalResource)
       return ((CanonicalResource) r).getName();
     return r.fhirType()+"/"+r.getId();
-  }
-
-  public class MyErrorListener implements ErrorListener {
-
-    @Override
-    public void error(TransformerException arg0) throws TransformerException {
-      System.out.println("XSLT Error: "+arg0.getMessage());
-      if (debug) {
-        arg0.printStackTrace();
-      }      
-    }
-
-    @Override
-    public void fatalError(TransformerException arg0) throws TransformerException {
-      System.out.println("XSLT Error: "+arg0.getMessage());
-      if (debug) {
-        arg0.printStackTrace();
-      }      
-    }
-
-    @Override
-    public void warning(TransformerException arg0) throws TransformerException {
-      System.out.println("XSLT Warning: "+arg0.getMessage());
-      if (debug) {
-        arg0.printStackTrace();
-      }      
-    }
-  }
-
-
-  private byte[] transform(byte[] source, byte[] xslt) throws TransformerException {
-    TransformerFactory f = TransformerFactory.newInstance();
-    f.setErrorListener(new MyErrorListener());
-    StreamSource xsrc = new StreamSource(new ByteArrayInputStream(xslt));
-    Transformer t = f.newTransformer(xsrc);
-    t.setErrorListener(new MyErrorListener());
-
-    StreamSource src = new StreamSource(new ByteArrayInputStream(source));
-    ByteArrayOutputStream out = new ByteArrayOutputStream();
-    StreamResult res = new StreamResult(out);
-    t.transform(src, res);
-    return out.toByteArray();
   }
 
   private Map<String, String> makeVars(FetchedResource r) {
