+++ org.hl7.fhir.publisher.core/src/test/java/tests/XSLTTransformerTest.java
@@ -1,0 +1,29 @@
+package tests;
+
+import org.apache.commons.io.IOUtils;
+import org.hl7.fhir.igtools.publisher.XSLTTransformer;
+import org.hl7.fhir.utilities.TextFile;
+import org.junit.jupiter.api.Test;
+
+import javax.xml.transform.TransformerException;
+import java.io.IOException;
+
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+
+public class XSLTTransformerTest {
+    @Test
+    public void test() throws IOException, TransformerException {
+        byte[] source = getLineSeparatorNormalizedBytes("/xslt/unicom-index.xml");
+        byte[] transform = getLineSeparatorNormalizedBytes("/xslt/unicom-transform.xslt");
+
+        XSLTTransformer xsltTransformer = new XSLTTransformer(false);
+        byte[] actual = xsltTransformer.transform(source, transform);
+        byte[] expected = getLineSeparatorNormalizedBytes("/xslt/unicom-expected.xml");
+
+        assertArrayEquals(actual, expected);
+    }
+
+    private byte[] getLineSeparatorNormalizedBytes(String fileName) throws IOException {
+        return new String(IOUtils.toByteArray(this.getClass().getResource(fileName))).replace(System.lineSeparator(), "\n").getBytes();
+    }
+}
